% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle}
\alias{mle}
\title{Compute MLE based on R1 and R2}
\usage{
mle(
  R1,
  R2,
  pi1,
  n,
  alpha1 = 0,
  alpha2 = 0,
  beta1 = 0,
  beta2 = 0,
  B = 100,
  seed = 18,
  eps = 10^(-6),
  iter_max = 15,
  gamma = 0.05,
  ...
)
}
\arguments{
\item{R1}{A \code{numeric} that provides the people of positive people in the sample that were known to be positive.}

\item{R2}{A \code{numeric} that provides the people of positive people in the sample.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha1}{A \code{numeric} that provides the False Negative (FN) rate for the sample R1. Default value is \code{0}.}

\item{alpha2}{A \code{numeric} that provides the False Negative (FN) rate for the sample R2. Default value is \code{0}.}

\item{beta1}{A \code{numeric} that provides the False Positive (FP) rate for the sample R1. Default value is \code{0}.}

\item{beta2}{A \code{numeric} that provides the False Positive (FP) rate for the sample R2. Default value is \code{0}.}

\item{B}{A \code{numeric} that corresponds to the number of simulations used in correction. This is only used in the case of
measurement error (i.e. \code{max(alpha1, alpha2, beta1, beta2) > 0}) in the simulation-based correction method.
Default value is \code{10^3}.}

\item{seed}{A \code{numeric} that provides the simulation seed. This is only used in the case of measurement error (i.e.
\code{max(alpha1, alpha2, beta1, beta2) > 0}) in the simulation-based correction method. Default value is \code{18}.}

\item{eps}{A \code{numeric} that corresponds to tolerance used to decide when to stop the simulation algorithm (iterative bootstrap).
This is only used in the case of measurement error (i.e. \code{max(alpha1, alpha2, beta1, beta2) > 0}) in the simulation-based correction method.
Default value is \code{10^(-6)}.}

\item{iter_max}{A \code{numeric} that corresponds to the maximum of iteration of the simulation algorithm (iterative bootstrap).
This is only used in the case of measurement error (i.e. \code{max(alpha1, alpha2, beta1, beta2) > 0}) in the simulation-based correction method.
Default value is \code{100}.}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item estimate: Estimated proportion
}
}
\description{
Proportion estimated using the MLE and confidence intervals based the asymptotic distribution of the estimator.
When measurement errors (false positive/negative) are considered, the estimator is corrected by simulations using the iterative
bootstrap.
}
\examples{
# Samples without measurement error
X = sim_Rs(pi2 = 3/100, pi1 = 1/100, n = 1500, seed = 18)
mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n)

# With measurement error
X = sim_Rs(pi2 = 30/1000, pi1 = 10/1000, n = 1500, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05, seed = 18)
mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n)
mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05)
}
\author{
Stephane Guerrier
}
