% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{modified_mle}
\alias{modified_mle}
\title{Compute approximated MLE based on R1 and R2}
\usage{
modified_mle(
  R1,
  R2,
  pi1,
  n,
  alpha1 = 0,
  alpha2 = 0,
  beta1 = 0,
  beta2 = 0,
  gamma = 0.05,
  ...
)
}
\arguments{
\item{R1}{A \code{numeric} that provides the people of positive people in the sample that were known to be positive.}

\item{R2}{A \code{numeric} that provides the people of positive people in the sample.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha1}{A \code{numeric} that provides the False Negative (FN) rate for the sample R1. Default value is \code{0}.}

\item{alpha2}{A \code{numeric} that provides the False Negative (FN) rate for the sample R2. Default value is \code{0}.}

\item{beta1}{A \code{numeric} that provides the False Positive (FP) rate for the sample R1. Default value is \code{0}.}

\item{beta2}{A \code{numeric} that provides the False Positive (FP) rate for the sample R2. Default value is \code{0}.}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item estimate: Estimated proportion
}
}
\description{
Proportion estimated using the approximated MLE and confidence intervals based the asymptotic distribution of the estimator or
the Clopper–Pearson approach when alpha2 = 0.
}
\examples{
# Samples without measurement error
X = sim_Rs(pi2 = 3/100, pi1 = 1/100, n = 1500, seed = 18)
modified_mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n)

# With measurement error
X = sim_Rs(pi2 = 30/1000, pi1 = 10/1000, n = 1500, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05, seed = 18)
modified_mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n)
modified_mle(R1 = X$R1, R2 = X$R2, pi1 = X$pi1, n = X$n, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05)
}
\author{
Stephane Guerrier
}
