% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{moment_estimator}
\alias{moment_estimator}
\title{Compute approximated MLE based on R0 and R}
\usage{
moment_estimator(
  R0,
  R,
  pi0,
  n,
  alpha0 = 0,
  alpha = 0,
  beta0 = 0,
  beta = 0,
  gamma = 0.05,
  ...
)
}
\arguments{
\item{R0}{A \code{numeric} that provides the people of positive people in the sample that were known to be positive.}

\item{R}{A \code{numeric} that provides the people of positive people in the sample.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{alpha}{A \code{numeric} that provides the False Negative (FN) rate for the sample R. Default value is \code{0}.}

\item{beta0}{A \code{numeric} that provides the False Positive (FP) rate for the sample R0. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item estimate: Estimated proportion
}
}
\description{
Proportion estimated using the approximated MLE and confidence intervals based the asymptotic distribution of the estimator or
the Clopper–Pearson approach when alpha = 0.
}
\examples{
# Samples without measurement error
X = sim_Rs(p = 3/100, pi0 = 1/100, n = 1500, seed = 18)
moment_estimator(R0 = X$R0, R = X$R, pi0 = X$pi0, n = X$n)

# With measurement error
X = sim_Rs(p = 30/1000, pi0 = 10/1000, n = 1500, alpha0 = 0.01,
alpha = 0.01, beta0 = 0.05, beta = 0.05, seed = 18)
moment_estimator(R0 = X$R0, R = X$R, pi0 = X$pi0, n = X$n)
moment_estimator(R0 = X$R0, R = X$R, pi0 = X$pi0, n = X$n, alpha0 = 0.01,
alpha = 0.01, beta0 = 0.05, beta = 0.05)
}
\author{
Stephane Guerrier
}
