% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{neg_log_like}
\alias{neg_log_like}
\title{Negative Log-Likelihood function based on R0 and R}
\usage{
neg_log_like(pi2, R, R0, pi0, n, alpha0, alpha, beta0, beta, ...)
}
\arguments{
\item{pi2}{A \code{numeric} value for p0.}

\item{R}{A \code{numeric} that provides the people of positive people in the sample.}

\item{R0}{A \code{numeric} that provides the people of positive people in the sample that were known to be positive.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{alpha}{A \code{numeric} that provides the False Negative (FN) rate for the sample R. Default value is \code{0}.}

\item{beta0}{A \code{numeric} that provides the False Positive (FP) rate for the sample R0. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{...}{Additional arguments.}
}
\value{
Negative Log-Likelihood
}
\description{
Log-Likelihood function based on R0 and R multiplied by -1.
}
\author{
Stephane Guerrier
}
