% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_a}
\alias{plot_a}
\title{Plot function a}
\usage{
plot_a(pi0, alpha, alpha0, beta, beta0, ...)
}
\arguments{
\item{pi0}{A \code{numeric} that provides the proportion of people (in the whole population) who are positive, as measured through a non-random,
but systematic sampling (e.g. based on medical selection).}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{...}{Additional arguments.}
}
\value{
A plot of the function a
}
\description{
Plot function a to assess the identifiability of the MLE
}
\examples{
plot_a(pi0 = 1/100, alpha0 = 0.005, alpha = 0.005, beta0 = 0.05, beta = 0.05)
}
\author{
Stephane Guerrier
}
