% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.CPreval}
\alias{print.CPreval}
\title{Print sample survey}
\usage{
\method{print}{CPreval}(x, ...)
}
\arguments{
\item{x}{A \code{CPreval} object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Prints the sample survey
}
\description{
Simple print function for the sample survey estimator
}
\examples{
X = sim_Rs(p = 3/100, pi0 = 1/100, n = 1500, seed = 18)
survey_sample(X$R, X$n)
}
\author{
Stephane Guerrier
}
\keyword{internal}
