% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.sample_survey}
\alias{print.sample_survey}
\title{Print sample survey}
\usage{
\method{print}{sample_survey}(x, ...)
}
\arguments{
\item{x}{A \code{sample_survey} object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Prints the sample survey
}
\description{
Simple print function for the sample survey estimator
}
\examples{
X = sim_Rs(p = 30/1000, pi0 = 10/1000, n = 1500, seed = 18)
survey_sample(X$R, X$n)
}
\author{
Stephane Guerrier
}
\keyword{internal}
