% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_Rs}
\alias{sim_Rs}
\title{Simulate R and R0}
\usage{
sim_Rs(p, pi0, n, alpha0 = 0, alpha = 0, beta0 = 0, beta = 0, seed = NULL, ...)
}
\arguments{
\item{p}{A \code{numeric} that provides the true poportion of proportion of people in this population who are positive.}

\item{pi0}{A \code{numeric} that provides the proportion of people (in the whole population) who are positive, as measured through a non-random,
but systematic sampling (e.g. based on medical selection).}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{seed}{A \code{numeric} that provides the simulation seed. Default value is \code{NULL}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} containing R, R0, pi0, n, alpha, alpha0, beta and beta0.
}
\description{
Simulation function for: R the number of people, out of n, that have been declared positive
in the random sample, and among these ones, the ones that have been declared as infected before, i.e. R0.
}
\examples{
# Samples without measurement error
sim_Rs(p = 3/100, pi0 = 1/100, n = 1500, seed = 18)

# With measurement error
sim_Rs(p = 3/100, pi0 = 1/100, n = 1500, alpha0 = 0.01,
alpha = 0.01, beta0 = 0.05, beta = 0.05, seed = 18)
}
\author{
Stephane Guerrier
}
