% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_Rs_fast}
\alias{sim_Rs_fast}
\title{Simulate R1 and R2 (fast version)}
\usage{
sim_Rs_fast(pi2, pi1, n, alpha1, alpha2, beta1, beta2, seed)
}
\arguments{
\item{pi2}{A \code{numeric} that provides the true poportion of proportion of people in this population who are positive.}

\item{pi1}{A \code{numeric} that provides the proportion of people (in the whole population) who are positive, as measured through a non-random,
but systematic sampling (e.g. based on medical selection).}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha1}{A \code{numeric} that provides the False Negative (FN) rate for the sample R1. Default value is \code{0}.}

\item{alpha2}{A \code{numeric} that provides the False Negative (FN) rate for the sample R2. Default value is \code{0}.}

\item{beta1}{A \code{numeric} that provides the False Positive (FP) rate for the sample R1. Default value is \code{0}.}

\item{beta2}{A \code{numeric} that provides the False Positive (FP) rate for the sample R2. Default value is \code{0}.}

\item{seed}{A \code{numeric} that provides the simulation seed. Default value is \code{NULL}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} containing R1 and R2
}
\description{
Faster simulation function for: R2 the number of people, out of n, that have been declared positive
in the random sample, and among these ones, the ones that have been declared as infected before, i.e. R1.
}
\examples{
# Samples without measurement error
sim_Rs_fast(pi2 = 3/100, pi1 = 1/100, n = 1500, seed = 18)

# With measurement error
sim_Rs_fast(pi2 = 3/100, pi1 = 1/100, n = 1500, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05, seed = 18)
}
\author{
Stephane Guerrier
}
