% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{survey_sample}
\alias{survey_sample}
\title{Compute proportion in the survey sample}
\usage{
survey_sample(R2, n, alpha2 = 0, beta2 = 0, gamma = 0.05, ...)
}
\arguments{
\item{R2}{A \code{numeric} that provides the people of positive people in the sample.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha2}{A \code{numeric} that provides the False Negative (FN) rate for the sample R2. Default value is \code{0}.}

\item{beta2}{A \code{numeric} that provides the False Positive (FP) rate for the sample R2. Default value is \code{0}.}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item estimate: Estimated proportion
 \item sd: Estimated standard error of the estimator
 \item ci: Confidence interval
 \item gamma: Confidence level (i.e. 1 - gamma) for confidence interval
}
}
\description{
Proportion estimated using the survey sample and confidence intervals based on the Clopper–Pearson approach.
}
\examples{
# Samples without measurement error
X = sim_Rs(pi2 = 30/1000, pi1 = 10/1000, n = 1500, seed = 18)
survey_sample(X$R2, X$n)

# With measurement error
X = sim_Rs(pi2 = 30/1000, pi1 = 10/1000, n = 1500, alpha1 = 0.01,
alpha2 = 0.01, beta1 = 0.05, beta2 = 0.05, seed = 18)
survey_sample(X$R2, X$n)
survey_sample(X$R2, X$n, alpha = 0.01, beta2 = 0.05)
}
\author{
Stephane Guerrier
}
