% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{survey_sample}
\alias{survey_sample}
\title{Compute proportion in the survey sample}
\usage{
survey_sample(R, n, alpha = 0, beta = 0, gamma = 0.05, ...)
}
\arguments{
\item{R}{A \code{numeric} that provides the people of positive people in the sample.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha}{A \code{numeric} that provides the False Negative (FN) rate for the sample R. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item estimate: Estimated proportion
 \item sd: Estimated standard error of the estimator
 \item ci_asy: Asymptotic confidence interval
 \item ci_cp: Confidence interval based on the Clopper–Pearson approach
 \item gamma: Confidence level (i.e. 1 - gamma) for confidence interval
 \item ...: Additional parameters
}
}
\description{
Proportion estimated using the survey sample and confidence intervals based on the Clopper–Pearson approach.
}
\examples{
# Samples without measurement error
X = sim_Rs(p = 30/1000, pi0 = 10/1000, n = 1500, seed = 18)
survey_sample(X$R, X$n)

# With measurement error
X = sim_Rs(p = 30/1000, pi0 = 10/1000, n = 1500, alpha0 = 0.01,
alpha = 0.01, beta0 = 0.05, beta = 0.05, seed = 18)
survey_sample(X$R, X$n)
survey_sample(X$R, X$n, alpha = 0.01, beta = 0.05)
}
\author{
Stephane Guerrier
}
